More unit test tweaks, and AV-1552 (partial), AV-2375, make goal test and cleanup
tools/cdk/include/rcc/rcc-worker.mk: put link of binary a BinLibLinks, not LibLinks
-- this is to separate the library links for generated code from binary code
-- part of AV-1552, and useful for ocpidev for proxies etc.
tools/cdk/include/test.mk: include script dependencies into Makefile, part of AV-2375
-- clean up .test goals into 3 categories: generic, global (shared) and local.
-- improving messaging
-- better error message when verifying without any executions, for all platforms or one platform
hdl/Makefile: ensure adapters are built before platforms (Adam)
tools/cdk/include/hdl/hdl-worker.mk: split binary links from generated links to support "generated"
-- AV-1552
tools/cdk/include/lib.mk: allow BuildImplementation and BuildModel to have a make goal
-- Do not consider .test directories without a Makefile
tools/cdk/include/libraries.mk: Use OcpiTestGoals for test goals
tools/cdk/include/library.mk: ditto, and add "generate" (AV-1552)
-- Only feed new test goals to .test directories with new Makefiles
tools/cdk/include/project.mk: ditto, and add hdladapters as a dependency to hdlplatforms
tools/cdk/include/util.mk: define the sharable list of test goals here (project, libraries, etc.)
tools/cdk/include/xxx-worker.mk: add "generate" for AV-1552, split links into genlinks and binlinks
tools/cdk/scripts/setsid.py: propagate ctrl-c to disassociated session ssh execution
tools/cdk/scripts/testrun.sh: allow for tput to not exist at all (embedded)
-- improve prettiness and indenting of message output
-- implement TestVerbose for putting everything on the console output
-- detect when there is no execution under "verifyonly"
tools/cdk/scripts/testrunprep.sh: improve prettiness and indenting of message output
tools/ocpigen/src/tests.cxx: improve prettiness and indenting of message output

+++ tools/cdk/scripts/testrun.sh
@@ -8,6 +8,8 @@
 # Ports are the names of output ports
 
 #echo PARMS: $*
+tput=tput
+[ -z "$(which tput)" ] && tput=true 
 spec=$1; shift
 component=${spec##*.}
 platform=$1; shift
@@ -50,29 +52,41 @@ function docase {
       [ ${#ports[@]} != 1 ] && output+="_$o"
       outputs+="$output=fileName=$3.$4.$2.$1.$o.out"
     done
-    echo Running $component test case: "$3.$4" on platform $platform using worker $2.$1... 1>&2
+    echo '  'Executing $component test case: "$3.$4" on platform $platform using worker $2.$1... 1>&2
     cmd=('OCPI_LIBRARY_PATH=../../../lib/rcc:../../gen/assemblies:$OCPI_CDK_DIR/lib/components/rcc' \
          ocpirun -d -v -m$component=$1 -w$component=$2 -P$component=$platform \
 	         --sim-dir=$3.$4.$2.$1.simulation \
 		 --dump-file=$3.$4.$2.$1.props $outputs ../../gen/applications/$3.$4.xml)
     rm -f -r $3.$4.$2.$1.*
-    if [ -z "$remote" -a -x runremote.sh ]; then
-      ./runremote.sh "(echo ${cmd[@]}; time ${cmd[@]})" > $3.$4.$2.$1.log 2>&1
+    if [ "$TestVerbose" = 1 ]; then
+	out=/dev/stdout
     else
-      (echo ${cmd[@]}; eval time env ${cmd[@]}) > $3.$4.$2.$1.log 2>&1 
+	out=/dev/null
+    fi
+    set -o pipefail
+    if [ -z "$remote" -a -x runremote.sh ]; then
+      ./runremote.sh "(echo ${cmd[@]}; time ${cmd[@]})" 2>&1 | tee $3.$4.$2.$1.log > $out
+    else
+      (echo ${cmd[@]}; eval time env ${cmd[@]}) 2>&1 | tee $3.$4.$2.$1.log > $out
     fi
     r=$?
+    set +o pipefail
     [ $r != 0 ] && {
-	tput setaf 1
+	$tput setaf 1
 	echo Execution FAILED - see log in run/$platform/$3.$4.$2.$1.log 1>&2
-	tput sgr0
+	$tput sgr0
 	failed=1
+	return 0
     }
   }
   [ -z "$view" -a -z "$verify" ] || 
     if [ "$r" = 0 ]; then
-      ../../gen/applications/verify_$3.sh $2.$1 $4 $view $verify
-      [ -n "$verify" -a $? = 0 -a "$KeepSimulations" != 1 ] && rm -r -f $3.$4.$2.$1.simulation
+      if [ -f $3.$4.$2.$1.props ]; then
+        ../../gen/applications/verify_$3.sh $2.$1 $4 $view $verify
+        [ -n "$verify" -a $? = 0 -a "$KeepSimulations" != 1 ] && rm -r -f $3.$4.$2.$1.simulation
+      else
+        echo No execution to verify for $3.$4 using $2.$1 on platform $platform. 1>&2 
+      fi
     else
       echo Execution failed so verify or view not performed.
     fi
